+++ implementation/src/main/java/com/avioconsulting/mule/testing/InvokerEventFactory.groovy
@@ -8,5 +8,9 @@ interface InvokerEventFactory {
 
     EventWrapper getMuleEventWithPayload(Object payload,
                                          String flowName,
+                                         String mediaType)
+
+    EventWrapper getMuleEventWithPayload(Object payload,
+                                         String flowName,
                                          Map attributes)
 }