+++ implementation/src/main/java/com/avioconsulting/mule/testing/dsl/invokers/JavaInvoker.groovy
@@ -2,4 +2,7 @@ package com.avioconsulting.mule.testing.dsl.invokers
 
 interface JavaInvoker {
     def inputPayload(inputObject)
+
+    def inputPayload(inputObject,
+                     String mediaType)
 }