+++ implementation/src/main/java/com/avioconsulting/mule/testing/dsl/invokers/JavaInvokerImpl.groovy
@@ -9,6 +9,7 @@ class JavaInvokerImpl implements
     private inputObject
     private final InvokerEventFactory eventFactory
     private final String flowName
+    private String mediaType
 
     JavaInvokerImpl(InvokerEventFactory eventFactory,
                     String flowName) {
@@ -20,9 +21,19 @@ class JavaInvokerImpl implements
         this.inputObject = inputObject
     }
 
+    @Override
+    def inputPayload(Object inputObject,
+                     String mediaType) {
+        this.inputObject = inputObject
+        this.mediaType = mediaType
+    }
+
     EventWrapper getEvent() {
-        eventFactory.getMuleEventWithPayload(inputObject,
-                                             flowName)
+        mediaType ? eventFactory.getMuleEventWithPayload(inputObject,
+                                                         flowName,
+                                                         mediaType) :
+                eventFactory.getMuleEventWithPayload(inputObject,
+                                                     flowName)
     }
 
     def transformOutput(EventWrapper event) {